// src/models/dto/hubtel/hb-enums.ts
export enum HbEnums {
    INITIATION = 'initiation',
    RESPONSE = 'response',
    TIMEOUTS = 'timeout',
    RELEASE = 'release',
    ADDTOCART = 'addtocart',

    DATATYPE_DISPLAY = 'display',
    DATATYPE_INPUT = 'display',

    FIELDTYPE_TEXT = 'text',
    FIELDTYPE_PHONE = 'phone',
    FIELDTYPE_NUMBER = 'number',
    FIELDTYPE_DECIMAL = 'decimal',

    ITEM_SEND_MONEY = 'Send Money'
}